python-manager/modules/converter-module/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
minio==7.2.0
pdf2docx==0.5.1
python-docx==0.8.11
requests==2.31.0
pydantic==2.5.0
